+++ src/ast-utils.js
@@ -1,17 +1,4 @@
-export function isReference(node, parent) {
-	if (parent.type === 'MemberExpression') return parent.computed || node === parent.object;
-
-	// disregard the `bar` in { bar: foo }
-	if (parent.type === 'Property' && node !== parent.value) return false;
-
-	// disregard the `bar` in `class Foo { bar () {...} }`
-	if (parent.type === 'MethodDefinition') return false;
-
-	// disregard the `bar` in `export { foo as bar }`
-	if (parent.type === 'ExportSpecifier' && node !== parent.local) return false;
-
-	return true;
-}
+export { default as isReference } from 'is-reference';
 
 export function flatten(node) {
 	const parts = [];