+++ test/form/multiple-var-declarations-b/output.js
@@ -1,5 +1,5 @@
 import './a';
-import a from 'commonjs-proxy:./a';
+import a from '_./a?commonjs-proxy';
 
 var b = 42;