+++ test/samples/custom-named-exports-browser-shims/main.js
@@ -1,0 +1,4 @@
+import { message, foo } from 'events';
+
+assert.equal( message, 'this is not builtin' );
+assert.equal( foo, 'this is a hidden export' );