+++ test/samples/custom-named-exports-warn-builtins/main.js
@@ -1,0 +1,3 @@
+import { message } from 'events';
+
+assert.equal( message, 'this is not builtin' );